+++ src/DataTransferObject/StationAddress.php
@@ -10,7 +10,7 @@ final class StationAddress
         public string $line1,
         public string $line2,
         public string $line3,
-        public string $line4,
+        public ?string $line4,
         public string $postcode,
     ) {
     }